Fix tickets not marked as paid.
